Add DsoCommand::configureWindow() unit tests
And fixed the bugs those tests revealed ;)

+++ src/cli/dsocommand.h
@@ -22,7 +22,7 @@ public slots:
 
 protected:
     AbstractPokitService * getService() override;
-    static quint32 maxWindowSize(const PokitProduct product);
+    static quint16 maxWindowSize(const PokitProduct product);
     static bool configureWindow(const PokitProduct product, const quint32 sampleRate, DsoService::Settings &settings);
 
 protected slots:
